+++ src/database/operations/index.ts
@@ -19,4 +19,5 @@ export * from './userDataProviderPermissions';
 export * from './userFunderPermissions';
 export * from './userGroupChangemakerPermissions';
 export * from './userGroupDataProviderPermissions';
+export * from './userGroupFunderPermissions';
 export * from './users';